Fix compile warnings

+++ src/util.cpp
@@ -81,7 +81,7 @@ bits64 hexadecimal_value_to_bits64(std::string str) {
     return get_bits64_from_vector(hexadecimal_str_to_binary_vector(str));
 }
 
-std::vector<bool> address_bits_from_address(int address, int tree_depth) {
+std::vector<bool> address_bits_from_address(int address, size_t tree_depth) {
     std::vector<bool> binary = convert_int_to_binary(address);
     std::vector<bool> result(tree_depth, 0);
